fastapi+sql/requirements.txt
fastapi
sqlalchemy
load_dotenv
PyJWT
bcrypt
uvicorn
